Tiny style tweaks to upstream socket abort PR
